// services/web-client/src/enums/error-messages.enum.ts
export enum ErrorMessages {
  REQUIRED_USERNAME_LOGIN = "Please enter your username.",
  REQUIRED_PASSWORD_LOGIN = "Please enter your password.",
  INVALID_CREDENTIALS_LOGIN = "Incorrect username or password. Please try again or reach out to the Admin.",
  INVALID_FULL_NAME = "Enter first and last name.",
  INVALID_PHONE = "Enter a valid phone number.",
  INVALID_ROLE = "Select a role.",
  REQUIRED_USERNAME = "Enter a unique username. Use 4–25 characters: letters and digits (0–9).",
  INVALID_USERNAME = "This username already exists in the system. Use 4–25 characters: letters and digits (0–9).",
  REQUIRED_PASSWORD = "Enter a password.",
  INVALID_PASSWORD = "Your password must meet all requirements.",
  INVALID_EMAIL = "Enter a valid email address.",
}